correct failed type mapping
float has been fine this whole time for scenes' cloud_cover, so
it's probably fine for score, which is a similar field

+++ app-backend/common/src/test/scala/com/implicits/Generators.scala
@@ -1150,7 +1150,7 @@ object Generators extends ArbitraryInstances {
       Gen.option(nonEmptyStringGen),
       Gen.const(true),
       Gen.option(uuidGen),
-      Gen.option(arbitrary[Int].map(_.toDouble))
+      Gen.option(arbitrary[Int].map(_.toFloat))
     ).mapN(AnnotationLabelWithClasses.Create.apply _)
 
   private def continentGen: Gen[Continent] =
